services/e14-service/requirements.txt
# E-14 Service Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
werkzeug==3.0.1
gunicorn==21.2.0
redis==5.0.1

# OCR and PDF Processing
anthropic==0.39.0
pdf2image==1.17.0
Pillow==10.4.0
PyMuPDF==1.25.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.1

# Utilities
prometheus-client==0.19.0
tenacity==8.2.3
